R plots

+++ bin/housekeeping.py
@@ -143,7 +143,9 @@ def clean_namespace(name):
     # remove uneeded files
     os.remove("clean.fasta")
     os.remove("ANARCI.txt")
-    os.remove("ab_contact.txt")
+
+    if os.path.exists("ab_contact.txt"):
+        os.remove("ab_contact.txt")
 
     os.rename(name + "_ANARCI_IMGT_annotated.fasta",
               name + "/" + name + "_ANARCI_IMGT_annotated.fasta")
